backend/requirements.txt
fastapi
uvicorn
python-chess
psycopg2-binary
python-dotenv
python-multipart
python-jose[cryptography]
bcrypt
requests
